fix(web): share chat
